+++ apps/wallet-mobile/src/features/Settings/useCases/changeAppSettings/ApplicationSettingsScreen.tsx
@@ -1,3 +1,4 @@
+import {networkConfigs} from '@yoroi/blockchains'
 import {isBoolean} from '@yoroi/common'
 import {SupportedThemes, useTheme} from '@yoroi/theme'
 import React from 'react'
@@ -14,7 +15,6 @@ import {defaultLanguage} from '../../../../kernel/i18n/languages'
 import {useCrashReports} from '../../../../yoroi-wallets/hooks'
 import {useAuthSetting, useAuthWithOs, useIsAuthOsSupported} from '../../../Auth/common/hooks'
 import {useSelectedNetwork} from '../../../WalletManager/common/hooks/useSelectedNetwork'
-import {networkConfigs} from '../../../WalletManager/network-manager/network-manager'
 import {useNavigateTo} from '../../common/navigation'
 import {SettingsSwitch} from '../../common/SettingsSwitch'
 import {NavigatedSettingsItem, SettingsItem, SettingsSection} from '../../SettingsItems'
