+++ apps/wallet-mobile/src/features/Settings/useCases/changeAppSettings/ChangeNetwork/NetworkTag.tsx
@@ -1,3 +1,4 @@
+import {networkConfigs} from '@yoroi/blockchains'
 import {useTheme} from '@yoroi/theme'
 import {Chain} from '@yoroi/types'
 import * as React from 'react'
@@ -11,7 +12,6 @@ import {useMetrics} from '../../../../../kernel/metrics/metricsManager'
 import {useWalletNavigation} from '../../../../../kernel/navigation'
 import {availableNetworks} from '../../../../WalletManager/common/constants'
 import {useWalletManager} from '../../../../WalletManager/context/WalletManagerProvider'
-import {networkConfigs} from '../../../../WalletManager/network-manager/network-manager'
 import {useStrings} from './strings'
 
 export const NetworkTag = ({
