+++ apps/wallet-mobile/src/features/Settings/useCases/changeWalletSettings/EasyConfirmation/EasyConfirmationScreen.stories.tsx
@@ -1,19 +1,20 @@
 import {storiesOf} from '@storybook/react-native'
+import {buildNetworkManagers} from '@yoroi/blockchains'
 import {Chain} from '@yoroi/types'
 import React from 'react'
 
+import {logger} from '../../../../../kernel/logger/logger'
 import {rootStorage} from '../../../../../kernel/storage/rootStorage'
 import {mocks} from '../../../../../yoroi-wallets/mocks/wallet'
 import {buildPortfolioTokenManagers} from '../../../../Portfolio/common/helpers/build-token-managers'
 import {WalletManagerProvider} from '../../../../WalletManager/context/WalletManagerProvider'
-import {buildNetworkManagers} from '../../../../WalletManager/network-manager/network-manager'
 import {WalletManager} from '../../../../WalletManager/wallet-manager'
 import {DisableEasyConfirmationScreen} from './DisableEasyConfirmationScreen'
 import {EnableEasyConfirmationScreen} from './EnableEasyConfirmationScreen'
 
 // TODO: should be mocked
 const {tokenManagers} = buildPortfolioTokenManagers()
-const networkManagers = buildNetworkManagers({tokenManagers})
+const networkManagers = buildNetworkManagers({tokenManagers, logger})
 const walletManager = new WalletManager({
   rootStorage,
   networkManagers,
