+++ apps/wallet-mobile/src/features/SetupWallet/legacy/SaveReadOnlyWallet/SaveReadOnlyWalletScreen.tsx
@@ -1,4 +1,5 @@
 import {useNavigation} from '@react-navigation/native'
+import {derivationConfig} from '@yoroi/blockchains'
 import {useAsyncStorage} from '@yoroi/common'
 import {Blockies} from '@yoroi/identicon'
 import {useSetupWallet} from '@yoroi/setup-wallet'
@@ -19,7 +20,6 @@ import {useMetrics} from '../../../../kernel/metrics/metricsManager'
 import {SetupWalletRouteNavigation} from '../../../../kernel/navigation'
 import {isEmptyString} from '../../../../kernel/utils'
 import {usePlate} from '../../../../yoroi-wallets/hooks'
-import {derivationConfig} from '../../../WalletManager/common/derivation-config'
 import {useCreateWalletXPub} from '../../../WalletManager/common/hooks/useCreateWalletXPub'
 import {useSelectedNetwork} from '../../../WalletManager/common/hooks/useSelectedNetwork'
 import {parseWalletMeta} from '../../../WalletManager/common/validators/wallet-meta'
