+++ apps/wallet-mobile/src/features/SetupWallet/useCases/RestoreHwWallet/SaveNanoXScreen.stories.tsx
@@ -1,19 +1,20 @@
 import {NavigationRouteContext} from '@react-navigation/native'
 import {action} from '@storybook/addon-actions'
 import {storiesOf} from '@storybook/react-native'
+import {buildNetworkManagers} from '@yoroi/blockchains'
 import {Api} from '@yoroi/types'
 import React from 'react'
 
+import {logger} from '../../../../kernel/logger/logger'
 import {rootStorage} from '../../../../kernel/storage/rootStorage'
 import {buildPortfolioTokenManagers} from '../../../Portfolio/common/helpers/build-token-managers'
 import {WalletManagerProvider} from '../../../WalletManager/context/WalletManagerProvider'
-import {buildNetworkManagers} from '../../../WalletManager/network-manager/network-manager'
 import {WalletManager} from '../../../WalletManager/wallet-manager'
 import {SaveNanoXScreen} from './SaveNanoXScreen'
 
 // TODO: should be mocked
 const {tokenManagers} = buildPortfolioTokenManagers()
-const networkManagers = buildNetworkManagers({tokenManagers})
+const networkManagers = buildNetworkManagers({tokenManagers, logger})
 const walletManager = new WalletManager({
   rootStorage,
   networkManagers,
