+++ apps/wallet-mobile/src/features/WalletManager/common/constants.ts
@@ -3,6 +3,7 @@ import {Chain, Wallet} from '@yoroi/types'
 import {freeze} from 'immer'
 
 import {isDev} from '../../../kernel/env'
+import {logger} from '../../../kernel/logger/logger'
 import {buildPortfolioTokenManagers} from '../../Portfolio/common/helpers/build-token-managers'
 
 export const addressModes: ReadonlyArray<Wallet.AddressMode> = freeze(['single', 'multiple'] as const)
@@ -12,8 +13,9 @@ export const implementations: ReadonlyArray<Wallet.Implementation> = freeze([
 ] as const)
 
 const {tokenManagers} = buildPortfolioTokenManagers()
-export const networkManagers = buildNetworkManagers({tokenManagers})
+export const networkManagers = buildNetworkManagers({tokenManagers, logger})
 
+// NOTE: needs update, SupportedNetworks is a client thing
 const supportedNetworksDev: Array<Chain.SupportedNetworks> = freeze([Chain.Network.Mainnet, Chain.Network.Preprod])
 
 const supportedNetworksProd: Array<Chain.SupportedNetworks> = freeze([Chain.Network.Mainnet, Chain.Network.Preprod])
