+++ apps/wallet-mobile/src/features/WalletManager/network-manager/get-wallet-factory.ts
@@ -1,7 +1,7 @@
-import {throwLoggedError} from '@yoroi/common'
 import {Chain, Wallet} from '@yoroi/types'
 import {freeze} from 'immer'
 
+import {throwLoggedError} from '../../../kernel/logger/helpers/throw-logged-error'
 import {makeCardanoWallet} from '../../../yoroi-wallets/cardano/cardano-wallet'
 import {networkManagers} from '../common/constants'
 import {WalletFactory} from '../common/types'
