+++ apps/wallet-mobile/src/kernel/logger/helpers/throw-logged-error.ts
@@ -1,0 +1,5 @@
+import {throwLoggedError as tle} from '@yoroi/common'
+
+import {logger} from '../logger'
+
+export const throwLoggedError: (error: string | Error) => never = tle(logger)
