+++ apps/wallet-mobile/src/yoroi-wallets/cardano/cardano-wallet.ts
@@ -4,6 +4,7 @@ import * as CSL from '@emurgo/cross-csl-core'
 import {createSignedLedgerTxFromCbor, signRawTransaction} from '@emurgo/yoroi-lib'
 import {Datum} from '@emurgo/yoroi-lib/dist/internals/models'
 import {AppApi} from '@yoroi/api'
+import {cardanoConfig, derivationConfig, protocolParamsPlaceholder} from '@yoroi/blockchains'
 import {isNonNullable} from '@yoroi/common'
 import {Api, App, Balance, HW, Network, Portfolio, Wallet} from '@yoroi/types'
 import {BigNumber} from 'bignumber.js'
@@ -17,9 +18,6 @@ import {toLedgerSignRequest} from '../../features/Discover/common/ledger'
 import {buildPortfolioBalanceManager} from '../../features/Portfolio/common/helpers/build-balance-manager'
 import {toBalanceManagerSyncArgs} from '../../features/Portfolio/common/transformers/toBalanceManagerSyncArgs'
 import {makeMemosManager, MemosManager} from '../../features/Transactions/common/memos/memosManager'
-import {cardanoConfig} from '../../features/WalletManager/common/adapters/cardano/cardano-config'
-import {derivationConfig} from '../../features/WalletManager/common/derivation-config'
-import {protocolParamsPlaceholder} from '../../features/WalletManager/network-manager/network-manager'
 import {LocalizableError} from '../../kernel/i18n/LocalizableError'
 import {throwLoggedError} from '../../kernel/logger/helpers/throw-logged-error'
 import {logger} from '../../kernel/logger/logger'
