+++ apps/wallet-mobile/src/yoroi-wallets/cardano/cip30/cip30-ledger.ts
@@ -2,10 +2,10 @@ import {MessageAddressFieldType, MessageData} from '@cardano-foundation/ledgerjs
 import {Transaction} from '@emurgo/cross-csl-core'
 import {createSignedLedgerTxFromCbor} from '@emurgo/yoroi-lib'
 import {normalizeToAddress} from '@emurgo/yoroi-lib/dist/internals/utils/addresses'
+import {cardanoConfig} from '@yoroi/blockchains'
 import {HW, Wallet} from '@yoroi/types'
 
 import {toLedgerSignRequest} from '../../../features/Discover/common/ledger'
-import {cardanoConfig} from '../../../features/WalletManager/common/adapters/cardano/cardano-config'
 import {buildCoseSign1FromSignature, makeCip8Key} from '../cip8/cip8'
 import {assertHasAllSigners} from '../common/signatureUtils'
 import {signMessageWithLedger, signTxWithLedger} from '../hw/hw'
