+++ apps/wallet-mobile/src/yoroi-wallets/cardano/hw/hw.ts
@@ -14,11 +14,10 @@ import AppAda, {DeviceStatusCodes} from '@cardano-foundation/ledgerjs-hw-app-car
 // @ts-ignore
 import TransportHID from '@emurgo/react-native-hid'
 import TransportBLE from '@ledgerhq/react-native-hw-transport-ble'
+import {cardanoConfig, derivationConfig} from '@yoroi/blockchains'
 import {HW, Wallet} from '@yoroi/types'
 import {BleError} from 'react-native-ble-plx'
 
-import {cardanoConfig} from '../../../features/WalletManager/common/adapters/cardano/cardano-config'
-import {derivationConfig} from '../../../features/WalletManager/common/derivation-config'
 import {ledgerMessages} from '../../../kernel/i18n/global-messages'
 import {LocalizableError} from '../../../kernel/i18n/LocalizableError'
 import {logger} from '../../../kernel/logger/logger'
