+++ apps/wallet-mobile/src/yoroi-wallets/cardano/usePrimaryTokenActivity.tsx
@@ -1,8 +1,8 @@
+import {networkConfigs} from '@yoroi/blockchains'
 import {fetchData, isRight} from '@yoroi/common'
 import {Chain} from '@yoroi/types'
 import {useQuery, UseQueryOptions} from 'react-query'
 
-import {networkConfigs} from '../../features/WalletManager/network-manager/network-manager'
 import {time} from '../../kernel/constants'
 import {queryInfo} from '../../kernel/query-client'
 import {CurrencySymbol, PriceMultipleResponse} from '../types/other'
