+++ apps/wallet-mobile/src/yoroi-wallets/mocks/WalletManagerProviderMock.tsx
@@ -1,17 +1,18 @@
+import {buildNetworkManagers} from '@yoroi/blockchains'
 import {Wallet} from '@yoroi/types'
 import * as React from 'react'
 
 import {buildPortfolioTokenManagers} from '../../features/Portfolio/common/helpers/build-token-managers'
 import {WalletManagerProvider} from '../../features/WalletManager/context/WalletManagerProvider'
-import {buildNetworkManagers} from '../../features/WalletManager/network-manager/network-manager'
 import {WalletManager} from '../../features/WalletManager/wallet-manager'
+import {logger} from '../../kernel/logger/logger'
 import {rootStorage} from '../../kernel/storage/rootStorage'
 import {YoroiWallet} from '../cardano/types'
 import {mocks} from './wallet'
 
 // TODO: should be mocked
 const {tokenManagers} = buildPortfolioTokenManagers()
-const networkManagers = buildNetworkManagers({tokenManagers})
+const networkManagers = buildNetworkManagers({tokenManagers, logger})
 export const walletManagerMock = new WalletManager({
   rootStorage,
   networkManagers,
