+++ apps/wallet-mobile/translations/messages/src/features/Settings/useCases/changeAppSettings/ApplicationSettingsScreen.json
@@ -6,12 +6,12 @@
     "start": {
       "line": 261,
       "column": 11,
-      "index": 9367
+      "index": 9331
     },
     "end": {
       "line": 264,
       "column": 3,
-      "index": 9477
+      "index": 9441
     }
   },
   {
@@ -21,12 +21,12 @@
     "start": {
       "line": 265,
       "column": 21,
-      "index": 9500
+      "index": 9464
     },
     "end": {
       "line": 268,
       "column": 3,
-      "index": 9633
+      "index": 9597
     }
   },
   {
@@ -36,12 +36,12 @@
     "start": {
       "line": 269,
       "column": 18,
-      "index": 9653
+      "index": 9617
     },
     "end": {
       "line": 272,
       "column": 3,
-      "index": 9765
+      "index": 9729
     }
   },
   {
@@ -51,12 +51,12 @@
     "start": {
       "line": 273,
       "column": 15,
-      "index": 9782
+      "index": 9746
     },
     "end": {
       "line": 276,
       "column": 3,
-      "index": 9888
+      "index": 9852
     }
   },
   {
@@ -66,12 +66,12 @@
     "start": {
       "line": 277,
       "column": 22,
-      "index": 9912
+      "index": 9876
     },
     "end": {
       "line": 280,
       "column": 3,
-      "index": 10033
+      "index": 9997
     }
   },
   {
@@ -81,12 +81,12 @@
     "start": {
       "line": 281,
       "column": 9,
-      "index": 10044
+      "index": 10008
     },
     "end": {
       "line": 284,
       "column": 3,
-      "index": 10144
+      "index": 10108
     }
   },
   {
@@ -96,12 +96,12 @@
     "start": {
       "line": 285,
       "column": 18,
-      "index": 10164
+      "index": 10128
     },
     "end": {
       "line": 288,
       "column": 3,
-      "index": 10291
+      "index": 10255
     }
   },
   {
@@ -111,12 +111,12 @@
     "start": {
       "line": 289,
       "column": 13,
-      "index": 10306
+      "index": 10270
     },
     "end": {
       "line": 292,
       "column": 3,
-      "index": 10415
+      "index": 10379
     }
   },
   {
@@ -126,12 +126,12 @@
     "start": {
       "line": 293,
       "column": 15,
-      "index": 10432
+      "index": 10396
     },
     "end": {
       "line": 296,
       "column": 3,
-      "index": 10545
+      "index": 10509
     }
   },
   {
@@ -141,12 +141,12 @@
     "start": {
       "line": 297,
       "column": 19,
-      "index": 10566
+      "index": 10530
     },
     "end": {
       "line": 300,
       "column": 3,
-      "index": 10727
+      "index": 10691
     }
   },
   {
@@ -156,12 +156,12 @@
     "start": {
       "line": 301,
       "column": 20,
-      "index": 10749
+      "index": 10713
     },
     "end": {
       "line": 304,
       "column": 3,
-      "index": 10883
+      "index": 10847
     }
   },
   {
@@ -171,12 +171,12 @@
     "start": {
       "line": 305,
       "column": 24,
-      "index": 10909
+      "index": 10873
     },
     "end": {
       "line": 308,
       "column": 3,
-      "index": 11092
+      "index": 11056
     }
   },
   {
@@ -186,12 +186,12 @@
     "start": {
       "line": 309,
       "column": 18,
-      "index": 11112
+      "index": 11076
     },
     "end": {
       "line": 312,
       "column": 3,
-      "index": 11243
+      "index": 11207
     }
   },
   {
@@ -201,12 +201,12 @@
     "start": {
       "line": 313,
       "column": 22,
-      "index": 11267
+      "index": 11231
     },
     "end": {
       "line": 316,
       "column": 3,
-      "index": 11448
+      "index": 11412
     }
   },
   {
@@ -216,12 +216,12 @@
     "start": {
       "line": 317,
       "column": 13,
-      "index": 11463
+      "index": 11427
     },
     "end": {
       "line": 320,
       "column": 3,
-      "index": 11571
+      "index": 11535
     }
   },
   {
@@ -231,12 +231,12 @@
     "start": {
       "line": 321,
       "column": 17,
-      "index": 11590
+      "index": 11554
     },
     "end": {
       "line": 324,
       "column": 3,
-      "index": 11707
+      "index": 11671
     }
   },
   {
@@ -246,12 +246,12 @@
     "start": {
       "line": 325,
       "column": 17,
-      "index": 11726
+      "index": 11690
     },
     "end": {
       "line": 328,
       "column": 3,
-      "index": 11849
+      "index": 11813
     }
   },
   {
@@ -261,12 +261,12 @@
     "start": {
       "line": 329,
       "column": 21,
-      "index": 11872
+      "index": 11836
     },
     "end": {
       "line": 333,
       "column": 3,
-      "index": 12090
+      "index": 12054
     }
   },
   {
@@ -276,12 +276,12 @@
     "start": {
       "line": 334,
       "column": 11,
-      "index": 12103
+      "index": 12067
     },
     "end": {
       "line": 337,
       "column": 3,
-      "index": 12207
+      "index": 12171
     }
   }
 ]
