+++ apps/wallet-mobile/translations/messages/src/features/SetupWallet/legacy/SaveReadOnlyWallet/SaveReadOnlyWalletScreen.json
@@ -6,12 +6,12 @@
     "start": {
       "line": 151,
       "column": 21,
-      "index": 4821
+      "index": 4792
     },
     "end": {
       "line": 154,
       "column": 3,
-      "index": 4948
+      "index": 4919
     }
   },
   {
@@ -21,12 +21,12 @@
     "start": {
       "line": 155,
       "column": 17,
-      "index": 4967
+      "index": 4938
     },
     "end": {
       "line": 158,
       "column": 3,
-      "index": 5081
+      "index": 5052
     }
   },
   {
@@ -36,12 +36,12 @@
     "start": {
       "line": 159,
       "column": 22,
-      "index": 5105
+      "index": 5076
     },
     "end": {
       "line": 162,
       "column": 3,
-      "index": 5229
+      "index": 5200
     }
   },
   {
@@ -51,12 +51,12 @@
     "start": {
       "line": 163,
       "column": 7,
-      "index": 5238
+      "index": 5209
     },
     "end": {
       "line": 166,
       "column": 3,
-      "index": 5336
+      "index": 5307
     }
   },
   {
@@ -66,12 +66,12 @@
     "start": {
       "line": 167,
       "column": 18,
-      "index": 5356
+      "index": 5327
     },
     "end": {
       "line": 170,
       "column": 3,
-      "index": 5477
+      "index": 5448
     }
   }
 ]
