+++ packages/blockchains/src/addresses/address-visual-derivation-path.ts
@@ -1,8 +1,8 @@
 import {Wallet} from '@yoroi/types'
 
-import {cardanoConfig} from '../../../features/WalletManager/common/adapters/cardano/cardano-config'
+import {cardanoConfig} from '../cardano/constants'
 
-export const derivationPathManagerMaker =
+export const addressVisualDerivationPathMaker =
   (implementation: Wallet.Implementation) =>
   ({account, role, index}: {account: number; role: number; index: number}) => {
     const {derivations} = cardanoConfig.implementations[implementation]
