+++ packages/blockchains/src/addresses/derivation-config.test.ts
@@ -1,0 +1,8 @@
+import {derivationConfig} from './derivation-config'
+
+describe('derivationConfig', () => {
+  // NOTE: while not in use within the package for coverage
+  it('should be defined', () => {
+    expect(derivationConfig).toBeDefined()
+  })
+})
