+++ packages/blockchains/src/cardano/helpers/epoch-progress.test.ts
@@ -1,6 +1,6 @@
 import {epochProgress} from './epoch-progress'
 import {dateToEpochInfo} from './date-to-epoch-info'
-import {networkConfigs} from '../../network-configs'
+import {networkConfigs} from '../../networks/network-configs'
 import {shelleyEraConfig} from '../constants'
 
 describe('epochProgress', () => {
