+++ packages/blockchains/src/index.ts
@@ -1,6 +1,10 @@
 export * from './types'
-export * from './network-manager'
-export * from './network-configs'
+
+export * from './addresses/address-visual-derivation-path'
+export * from './addresses/derivation-config'
+
+export * from './networks/network-manager'
+export * from './networks/network-configs'
 
 export * from './cardano/constants'
 export * from './cardano/helpers/date-to-epoch-info'
