+++ packages/blockchains/src/networks/network-configs.test.ts
